Update admin panel to edit post title/body

+++ public/js/create.ng.js
@@ -18,6 +18,19 @@ create.controller('createCtrl', ['$scope', '$http', '$window',
                 console.log('Error: ' + data);
             });
 
+        // format date function, identical to the one in qcms.ng.js
+        $scope.formatDate = function(dateStr) {
+            var date = new Date(dateStr);
+            var hours = date.getHours();
+            var mins = date.getMinutes();
+            if (mins < 10) mins = '0' + mins;
+            var meridian = 'AM';
+            if (hours >= 12) meridian = 'PM';
+            if (hours > 12) hours -= 12;
+            var display = date.toLocaleDateString() + ', ' + hours + ':' + mins + ' ' + meridian;
+            return display;
+        };
+
         // submitting form, send post to node API
         $scope.createPost = function() {
             var title = $scope.formData.title;
@@ -35,15 +48,41 @@ create.controller('createCtrl', ['$scope', '$http', '$window',
             // otherwise do nothing, not long enough
         };
 
-        // delete post when checked
-        $scope.deletePost = function(id) {
-            $http.delete('/post/' + id)
-                .success(function(data) {
-                    $scope.posts = data;
-                })
-                .error(function(data) {
-                    console.log('Error: ' + data);
-                });
+        // edit post
+        $scope.editPost = function(id) {
+            var confirm = $window.confirm("Note that edited posts should have correct HTML, as the server will save as is. Confirm?");
+            // if confirm, actually send request to server to edit post
+            if (confirm) {
+                // find correct index
+                var index = -1;
+                for (var i = 0; i < $scope.posts.length; i++) {
+                    if ($scope.posts[i]._id === id) {
+                        index = i;
+                    }
+                }
+                $http.post('/post/' + id, $scope.posts[index])
+                    .success(function(data) {
+                        $window.alert("Updated successful");
+                    })
+                    .error(function(err) {
+                        console.log('Error occurred editing post: ' + err);
+                    });
+            }
+        }
+
+        // delete post
+        $scope.deletePost = function(id, title) {
+            var confirm = $window.confirm("Deleting post titled '" + title + "', confirm?");
+            // if 'ok' clicked, actually send request to server to delete post
+            if (confirm) {
+                $http.delete('/post/' + id)
+                    .success(function(data) {
+                        $scope.posts = data;
+                    })
+                    .error(function(err) {
+                        console.log('Error occurred deleting post: ' + err);
+                    });
+            }
         };
 
         // login
